+++ CHANGELOG.md
@@ -6,6 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- [@http/html-stream] `partialHtml` token removed, regular `safe` token used
+  instead
+
+### Added
+
+- [@http/html-stream/transform] `tokenize` tokenizes `safe` chunks of html into
+  tag tokens
+
 ## [0.9.0]
 
 ### Changed